tests: hold a code address in a fixed-width integer in the static-init fixture
`unsigned long` is 32 bits on LLP64, where the cast of a block address would
truncate. The Windows fixture lists compile this one.

+++ tests/fixtures/c/static_init_logical_and.c
@@ -22,7 +22,8 @@ static int classify(int n) {
     return (int)table[n].value;
 }
 
-typedef unsigned long word;
+// Wide enough to hold a code address on every supported data model.
+typedef unsigned long long word;
 
 static int dispatch(int n) {
     // The block-address form in the same shape: prefix `&&`, runtime stores.
